Add the metashape deployment files
NOTE: Running metashape in pods for use in the MDDP virtual desktop is
currently broken until monash-01 to monash-02 network issues are solved.
Running the license server in a pod looks like it will also not work as
it sees the host as different each time and the keys as invalid.