Moving more setup to fluxcd and rework asdc_run/webapp init
- More config and secrets moved to flu-configmap.yaml
- WebODM init script now defined in asdc_init.sh and loaded via
configmap, making it easier to modify and reapply
-SSL setup done in the webapp-worker pod by above script, faster and
much more self-contained. Certs are stored on the media drive now so do
not need to be copied in and out.
- NodeODM setup script now runs node server too
- Removed floating ip setup/config, not currently used but kept in
floating_ip_setup.sh, leaning towards manual fip creation for new
deployments instead of attempting to automate a process run so rarely.