Update db_check script logic and fix an error (no user)
The disk filling problem ocurred again, the check detected this, but
the fix failed as we forgot the postgres user for the vacuum command.
This is now corrected and also changed logic so if the fix fails to
free disk space, instead of continuing as usual we sleep to allow
manual fixes to be applied rather than enter a restart loop.